Add base storage class
