+++ parsec/core/fs/storage/__init__.py
@@ -1,11 +1,13 @@
 # Parsec Cloud (https://parsec.cloud) Copyright (c) AGPLv3 2019 Scille SAS
 
+from parsec.core.fs.storage.base_storage import BaseStorage
 from parsec.core.fs.storage.user_storage import UserStorage
 from parsec.core.fs.storage.manifest_storage import ManifestStorage
 from parsec.core.fs.storage.chunk_storage import ChunkStorage, BlockStorage
 from parsec.core.fs.storage.workspace_storage import WorkspaceStorage, WorkspaceStorageTimestamped
 
 __all__ = (
+    "BaseStorage",
     "ManifestStorage",
     "ChunkStorage",
     "BlockStorage",
